+++ CHANGELOG.md
@@ -52,7 +52,7 @@
 Previously required a `type => matcher` map. Options are checked against `type => matcher` coercion input, and a
 descriptive error is thrown when api is created with the old options format.
 
-* **BREAKING**: Renamed `middlewares` to `middleware` and `:middlewares` metahandler to `:middleware`
+* **BREAKING**: Renamed `middlewares` to `middleware` and `:middlewares` key (restructuring) to `:middleware`
 
 * **BREAKING**: Middleware must be defined as data: both our forms take a vector
 of middleware containing either a) fully configured mws (function)